requirements.txt
opencv-python>=4.10,<5
imagehash>=4.3.1,<5
pillow>=10.4.0,<11
